odroid: defconfig: configure keyed autoboot
At least the ODROID-X2 is afflicted by the issue, that if nothing
is connected to the UART port during boot, random data is produced
which then halts u-boot's autoboot process.

Hence we configure keyed autoboot with the sequence <ESC><ESC>,
which has a low probability of showing up in the aforementioned
random data.